added_angular

+++ src/main/java/lee/projects/music/UserResource.java
@@ -1,5 +1,7 @@
 package lee.projects.music;
 
+import lee.projects.music.exceptions.IdExistsException;
+import lee.projects.music.exceptions.UserNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -11,10 +13,16 @@ import java.util.List;
 public class UserResource {
 
     private final UserService userService;
+    private final UserRepository userRepository;
 
-    public UserResource(UserService userService) {
+    public UserResource(UserService userService, UserRepository userRepository) {
         this.userService = userService;
+        this.userRepository = userRepository;
     }
+    public boolean idExists(Long id) {
+        return userRepository.existsById(id);
+    }
+
     @GetMapping("/all")
     public ResponseEntity<List<User>> getAllUsers(){
         List<User> users = userService.findAllUsers();
@@ -23,8 +31,12 @@ public class UserResource {
     @GetMapping("/find/{id}")
     public ResponseEntity<User> getUserById(@PathVariable("id") Long id){
         User user = userService.findUserById(id);
+        if(!idExists(id)){
+            throw new UserNotFoundException("ID no. " + id + " not found");
+        }
         return new ResponseEntity<>(user, HttpStatus.OK);
     }
+
     @PostMapping("/add")
     public ResponseEntity<User> addUser(@RequestBody User user){
         User newUser = userService.addUser(user);
@@ -36,8 +48,11 @@ public class UserResource {
         return new ResponseEntity<>(updateUser, HttpStatus.CREATED);
     }
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<?> deleteUser(@PathVariable("id") Long id){
+    public ResponseEntity<?> deleteUser(@PathVariable("id") Long id) throws IdExistsException {
+        if(!idExists(id)){
+            throw new IdExistsException("ID no. " + id + " not found");
+        }
         userService.deleteUser(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+        return new ResponseEntity<>("Id no. " + id + " successfully deleted", HttpStatus.OK);
     }
 }
