+++ src/main/java/lee/projects/music/UserService.java
@@ -1,7 +1,9 @@
 package lee.projects.music;
 
+import lee.projects.music.exceptions.IdExistsException;
 import lee.projects.music.exceptions.UserNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,21 +16,31 @@ public class UserService {
     public UserService(UserRepository userRepository) {
         this.userRepository = userRepository;
     }
+    public boolean idExists(Long id) {
+        return userRepository.existsById(id);
+    }
 
-    public User addUser(User user){
+    public User addUser(User user) {
         return userRepository.save(user);
     }
-    public List<User> findAllUsers(){
+
+    public List<User> findAllUsers() {
         return userRepository.findAll();
     }
-    public User updateUser(User user){
+
+    public User updateUser(User user) {
         return userRepository.save(user);
     }
-    public User findUserById(Long id){
+
+    public User findUserById(Long id) {
         return userRepository.findUserById(id)
                 .orElseThrow(() -> new UserNotFoundException("User by ID " + id + " not found"));
     }
-    public void deleteUser(Long id){
+
+    public void deleteUser(Long id) {
+        if (!idExists(id)) {
+            throw new IdExistsException("ID no. " + id + " not found");
+        }
         userRepository.deleteUserById(id);
     }
 }
